# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: NO-Houchi-Bicycle-Net API (prototype draft)
  version: 0.2.0
  description: >
    試作品スコープ向けの API 草案。通報、持ち主による解除、未解除時の回収依頼、
    回収結果記録を中心に扱う。ブラックリスト、高度分析、外部連携、詳細インセンティブは将来構想。
servers:
  - url: http://localhost:3000
paths:
  /api/auth/login:
    post:
      summary: 管理者/サポーターのログイン
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/reports:
    get:
      summary: 管理画面向け通報一覧
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ReportStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BicycleReport"
    post:
      summary: Androidアプリから通報を作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BicycleReport"
  /api/reports/{id}:
    get:
      summary: 通報詳細
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BicycleReport"
  /api/reports/{id}/collection-request:
    post:
      summary: 未解除案件を回収依頼対象にする
      parameters:
        - $ref: "#/components/parameters/ReportId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionRequestBody"
      responses:
        "200":
          description: Collection requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BicycleReport"
  /api/reports/{id}/collection-result:
    patch:
      summary: 回収結果を記録する
      parameters:
        - $ref: "#/components/parameters/ReportId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionResultBody"
      responses:
        "200":
          description: Collection result recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BicycleReport"
  /api/owner/markers/{code}:
    get:
      summary: 持ち主向けの警告内容取得
      parameters:
        - $ref: "#/components/parameters/MarkerCode"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerMarkerResponse"
        "404":
          description: Marker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/owner/markers/{code}/unlock-temp:
    post:
      summary: 持ち主による仮解除
      parameters:
        - $ref: "#/components/parameters/MarkerCode"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockTempBody"
      responses:
        "200":
          description: Temporary unlock created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnlockTempResponse"
  /api/owner/markers/{code}/unlock-final:
    post:
      summary: 持ち主による本解除
      parameters:
        - $ref: "#/components/parameters/MarkerCode"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockFinalBody"
      responses:
        "200":
          description: Final unlock completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnlockFinalResponse"
components:
  parameters:
    ReportId:
      name: id
      in: path
      required: true
      schema:
        type: string
    MarkerCode:
      name: code
      in: path
      required: true
      schema:
        type: string
  schemas:
    ReportStatus:
      type: string
      enum:
        - reported
        - temporary
        - resolved
        - collection_requested
        - collected
        - not_found_on_collection
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
    ReportCreate:
      type: object
      required: [imageUrl, latitude, longitude, markerCode, identifierText]
      properties:
        imageUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        markerCode:
          type: string
        identifierText:
          type: string
        notes:
          type: string
    BicycleReport:
      type: object
      properties:
        id:
          type: string
        markerId:
          type: string
        imageUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
          nullable: true
        identifierText:
          type: string
        status:
          $ref: "#/components/schemas/ReportStatus"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CollectionRequestBody:
      type: object
      properties:
        notes:
          type: string
        requestedBy:
          type: string
    CollectionResultBody:
      type: object
      required: [result]
      properties:
        result:
          type: string
          enum: [collected, not_found_on_collection]
        notes:
          type: string
        resultRecordedBy:
          type: string
    OwnerMarkerResponse:
      type: object
      properties:
        marker:
          type: object
          properties:
            code:
              type: string
        report:
          allOf:
            - $ref: "#/components/schemas/BicycleReport"
          nullable: true
        declaration:
          $ref: "#/components/schemas/Declaration"
    Declaration:
      type: object
      properties:
        id:
          type: string
        markerId:
          type: string
        declaredAt:
          type: string
          format: date-time
        eligibleFinalAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        finalizedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      nullable: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          nullable: true
    UnlockTempBody:
      type: object
      properties:
        notes:
          type: string
    UnlockTempResponse:
      type: object
      properties:
        declaredAt:
          type: string
          format: date-time
        eligibleFinalAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [temporary]
    UnlockFinalBody:
      type: object
      required:
        - scannedCode
      properties:
        scannedCode:
          type: string
        ownerEmail:
          type: string
          format: email
    UnlockFinalResponse:
      type: object
      properties:
        finalizedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [resolved]
        message:
          type: string
        coupon:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
            description:
              type: string
            shopName:
              type: string
            discount:
              type: number
            discountType:
              type: string
            expiresAt:
              type: string
              format: date-time